+++ pymt_heatf/__init__.py
@@ -4,8 +4,8 @@ import pkg_resources
 __version__ = pkg_resources.get_distribution("pymt_heatf").version
 
 
-from .bmi import HeatBMI
+from .bmi import HeatModelF
 
 __all__ = [
-    "HeatBMI",
+    "HeatModelF",
 ]